Выносит логику в контроллер
